third_party/WebKit/LayoutTests/fast/pagination/caret-range-outside-paged-x-rtl-vertical-rl-expected.txt

Test that hitting areas outside a paged container takes us to the right text content on some page.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
Below page 1:
PASS document.caretRangeFromPoint(499,320).startOffset is 0
PASS document.caretRangeFromPoint(430,320).startOffset is 0
PASS document.caretRangeFromPoint(390,320).startOffset is 4
Above page 1:
PASS document.caretRangeFromPoint(499,0).startOffset is 0
PASS document.caretRangeFromPoint(430,0).startOffset is 3
PASS document.caretRangeFromPoint(390,0).startOffset is 7
To the right of page 1:
PASS document.caretRangeFromPoint(499,320).startOffset is 0
PASS document.caretRangeFromPoint(499,250).startOffset is 0
PASS document.caretRangeFromPoint(499,0).startOffset is 0
Scroll to page 2
Below page 2:
PASS document.caretRangeFromPoint(449,320).startOffset is 12
PASS document.caretRangeFromPoint(330,320).startOffset is 20
Above page 2:
PASS document.caretRangeFromPoint(449,0).startOffset is 15
PASS document.caretRangeFromPoint(330,0).startOffset is 23
Scroll to page 3
Below page 3:
PASS document.caretRangeFromPoint(449,320).startOffset is 24
PASS document.caretRangeFromPoint(330,320).startOffset is 32
Above page 3:
PASS document.caretRangeFromPoint(449,0).startOffset is 27
PASS document.caretRangeFromPoint(330,0).startOffset is 35
xxx xxx xxx xxx xxx xxx xxx xxx xxx
